+++ src/main/java/focandlol/recipeproject/global/exception/ErrorCode.java
@@ -26,6 +26,7 @@ public enum ErrorCode {
    * 시스템 에러
    */
   INTERNAL_SERVER_ERROR("서버 오류입니다.", HttpStatus.INTERNAL_SERVER_ERROR),
+  METHOD_ARGUMENT_NOT_VALID("잘못된 입력 데이터입니다", BAD_REQUEST),
 
   INVALID_TAG("존재하지 않는 태그입니다.",BAD_REQUEST),
   EXIST_TAG("이미 존재하는 태그입니다",BAD_REQUEST),
