+++ src/main/java/focandlol/recipeproject/tag/controller/TagController.java
@@ -1,8 +1,10 @@
 package focandlol.recipeproject.tag.controller;
 
+import focandlol.recipeproject.tag.dto.CreateTagDto;
 import focandlol.recipeproject.tag.dto.TagDto;
 import focandlol.recipeproject.tag.dto.UpdateTagDto;
 import focandlol.recipeproject.tag.service.TagService;
+import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,8 +20,8 @@ public class TagController {
   private final TagService tagService;
 
   @PostMapping("/tag")
-  public List<TagDto> addTag(@RequestBody List<String> tags) {
-    return tagService.add(tags);
+  public List<TagDto> addTag(@RequestBody @Valid CreateTagDto createTagDto) {
+    return tagService.add(createTagDto.getTags());
   }
 
   @DeleteMapping("/tag")
@@ -28,7 +30,7 @@ public class TagController {
   }
 
   @PutMapping("/tag")
-  public void updateTag(@RequestBody UpdateTagDto updateTagDto) {
+  public void updateTag(@RequestBody @Valid UpdateTagDto updateTagDto) {
     tagService.update(updateTagDto.getTag(), updateTagDto.getChange());
   }
 }
