+++ src/main/java/focandlol/recipeproject/tag/dto/CreateTagDto.java
@@ -1,7 +1,9 @@
 package focandlol.recipeproject.tag.dto;
 
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -9,12 +11,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateTagDto {
+public class CreateTagDto {
 
-  @NotBlank
-  private String tag;
-
-  @NotBlank
-  @Size(max = 20)
-  private String change;
+  @NotEmpty
+  List<@NotBlank @Size(max = 20) String> tags;
 }
